did some refactoring and worked on unit testing

+++ its-a-feature-not-a-bug/app/src/test/java/com/example/its_a_feature_not_a_bug/AddEventTest.java
@@ -10,6 +10,6 @@ public class AddEventTest {
         assertEquals("Test Event", event.getTitle());
         assertEquals("Test Host", event.getHost());
         assertEquals("Test Description", event.getDescription());
-        assertEquals(10, event.getAttendeeLimit());
+        assertEquals(10, (int) event.getAttendeeLimit());
     }
 }
